ChangeList: C compiler performance has been significantly improved when compiling large and complex source files.
Notes: This was done by eliminating linear search of the expression
tree in FindFuncStmtTree() in cgen2.c. Instead, cstmt2.c will now 
store a pointer to the tree node in GenFunctionNode(). Then we 
can use SYM_HANDLE lookup (which is meant to be fast) instead of 
the far slower linear search.

P4:26354